+++ pkg/scalers/cpu_memory_scaler_test.go
@@ -6,12 +6,12 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/stretchr/testify/assert"
-	"k8s.io/api/autoscaling/v2beta2"
+	v2 "k8s.io/api/autoscaling/v2"
 	v1 "k8s.io/api/core/v1"
 )
 
 type parseCPUMemoryMetadataTestData struct {
-	metricType v2beta2.MetricTargetType
+	metricType v2.MetricTargetType
 	metadata   map[string]string
 	isError    bool
 }
@@ -32,11 +32,11 @@ var testCPUMemoryMetadata = []parseCPUMemoryMetadataTestData{
 	{"", validCPUMemoryMetadata, false},
 	{"", validContainerCPUMemoryMetadata, false},
 	{"", map[string]string{"type": "Utilization", "value": "50"}, false},
-	{v2beta2.UtilizationMetricType, map[string]string{"value": "50"}, false},
+	{v2.UtilizationMetricType, map[string]string{"value": "50"}, false},
 	{"", map[string]string{"type": "AverageValue", "value": "50"}, false},
-	{v2beta2.AverageValueMetricType, map[string]string{"value": "50"}, false},
+	{v2.AverageValueMetricType, map[string]string{"value": "50"}, false},
 	{"", map[string]string{"type": "Value", "value": "50"}, true},
-	{v2beta2.ValueMetricType, map[string]string{"value": "50"}, true},
+	{v2.ValueMetricType, map[string]string{"value": "50"}, true},
 	{"", map[string]string{"type": "AverageValue"}, true},
 	{"", map[string]string{"type": "xxx", "value": "50"}, true},
 }
@@ -65,21 +65,21 @@ func TestGetMetricSpecForScaling(t *testing.T) {
 	scaler, _ := NewCPUMemoryScaler(v1.ResourceCPU, config)
 	metricSpec := scaler.GetMetricSpecForScaling(context.Background())
 
-	assert.Equal(t, metricSpec[0].Type, v2beta2.ResourceMetricSourceType)
+	assert.Equal(t, metricSpec[0].Type, v2.ResourceMetricSourceType)
 	assert.Equal(t, metricSpec[0].Resource.Name, v1.ResourceCPU)
-	assert.Equal(t, metricSpec[0].Resource.Target.Type, v2beta2.UtilizationMetricType)
+	assert.Equal(t, metricSpec[0].Resource.Target.Type, v2.UtilizationMetricType)
 
 	// Using trigger.metricType field for type
 	config = &ScalerConfig{
 		TriggerMetadata: map[string]string{"value": "50"},
-		MetricType:      v2beta2.UtilizationMetricType,
+		MetricType:      v2.UtilizationMetricType,
 	}
 	scaler, _ = NewCPUMemoryScaler(v1.ResourceCPU, config)
 	metricSpec = scaler.GetMetricSpecForScaling(context.Background())
 
-	assert.Equal(t, metricSpec[0].Type, v2beta2.ResourceMetricSourceType)
+	assert.Equal(t, metricSpec[0].Type, v2.ResourceMetricSourceType)
 	assert.Equal(t, metricSpec[0].Resource.Name, v1.ResourceCPU)
-	assert.Equal(t, metricSpec[0].Resource.Target.Type, v2beta2.UtilizationMetricType)
+	assert.Equal(t, metricSpec[0].Resource.Target.Type, v2.UtilizationMetricType)
 }
 
 func TestGetContainerMetricSpecForScaling(t *testing.T) {
@@ -90,21 +90,21 @@ func TestGetContainerMetricSpecForScaling(t *testing.T) {
 	scaler, _ := NewCPUMemoryScaler(v1.ResourceCPU, config)
 	metricSpec := scaler.GetMetricSpecForScaling(context.Background())
 
-	assert.Equal(t, metricSpec[0].Type, v2beta2.ContainerResourceMetricSourceType)
+	assert.Equal(t, metricSpec[0].Type, v2.ContainerResourceMetricSourceType)
 	assert.Equal(t, metricSpec[0].ContainerResource.Name, v1.ResourceCPU)
-	assert.Equal(t, metricSpec[0].ContainerResource.Target.Type, v2beta2.UtilizationMetricType)
+	assert.Equal(t, metricSpec[0].ContainerResource.Target.Type, v2.UtilizationMetricType)
 	assert.Equal(t, metricSpec[0].ContainerResource.Container, validContainerCPUMemoryMetadata["containerName"])
 
 	// Using trigger.metricType field for type
 	config = &ScalerConfig{
 		TriggerMetadata: map[string]string{"value": "50", "containerName": "bar"},
-		MetricType:      v2beta2.UtilizationMetricType,
+		MetricType:      v2.UtilizationMetricType,
 	}
 	scaler, _ = NewCPUMemoryScaler(v1.ResourceCPU, config)
 	metricSpec = scaler.GetMetricSpecForScaling(context.Background())
 
-	assert.Equal(t, metricSpec[0].Type, v2beta2.ContainerResourceMetricSourceType)
+	assert.Equal(t, metricSpec[0].Type, v2.ContainerResourceMetricSourceType)
 	assert.Equal(t, metricSpec[0].ContainerResource.Name, v1.ResourceCPU)
-	assert.Equal(t, metricSpec[0].ContainerResource.Target.Type, v2beta2.UtilizationMetricType)
+	assert.Equal(t, metricSpec[0].ContainerResource.Target.Type, v2.UtilizationMetricType)
 	assert.Equal(t, metricSpec[0].ContainerResource.Container, "bar")
 }